-- moved the root route / of the app to the home page -- moved quiz page to /live route
