--- webapp/brilla-frontend/app/home/page.tsx
@@ -1,18 +1,0 @@
-import RootLayout from "../layout";
-import Navbar from "@/components/navbar";
-import HomeBase from "@/components/home-page";
-import Footer from "@/components/footer";
-
-const HomePage = () => {
-  return (
-    <RootLayout>
-      <main className="flex flex-col">
-      <Navbar gradientBg={false} />
-      <HomeBase />
-      <Footer />
-      </main>
-    </RootLayout>
-  );
-};
-
-export default HomePage;
